feat(cmd): create ggshield sca scan ci command
Issue: SCA-853
