+++ ggshield/cmd/common_options.py
@@ -202,3 +202,10 @@ directory_argument = click.argument(
     type=click.Path(exists=True, readable=True, path_type=Path, file_okay=False),
     required=False,
 )
+
+all_option = click.option(
+    "--all",
+    is_flag=True,
+    default=None,
+    help="Report all vulnerabilities in the final state.",
+)
