+++ ggshield/cmd/sca/scan/scan_common_options.py
